Create WebDev5.html
FInished adding in a feature layer of bus shelters in the MSP metro area. Also changed the title so when it is loading it says metro bus shelters, then I added in the URL from arc online and had to add in a /0 at the end so it knew  its position among the other potential layers. Sorry I didnt make a seperate commit for each step kinda did it all at once.  SORRY!!!